requirements.txt
fastapi
uvicorn
langchain
langchain-community
sentence-transformers
faiss-cpu
pypdf
python-multipart
groq
